+++ src/main/java/com/nylas/GrantQuery.java
@@ -1,0 +1,5 @@
+package com.nylas;
+
+public class GrantQuery extends RestfulQuery<GrantQuery> {
+
+}