+++ src/main/java/com/nylas/IntegrationQuery.java
@@ -1,0 +1,5 @@
+package com.nylas;
+
+public class IntegrationQuery extends RestfulQuery<IntegrationQuery> {
+
+}