+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+- Added UAS Support (Integrations API, Grants API, UAS Hosted Authentication)
+
 ### Changed
 
 - Bump supported API version to v2.5